Move hit component to new file
